put event firebase code in db class
